refactor(#75): simplify threads implementation
- Replace fmt.Printf with output.PrintSummary in displayThreads (correct stream/behavior)
- Move ParseODataDuration and ThreadDuration JSON tests to model package
- Remove parseODataDuration wrapper from cmd package
- Fix err scoping in duration parse block
